OFJ-1824 As an administrator I should be able to define the max upload file size
1. Fix several typo
2. Change the column length of max_upload_file_size.
3. Change the validator accordingly.

+++ library/Fisma/String.php
@@ -283,7 +283,7 @@ class Fisma_String
     public static function convertFilesizeToInteger($value)
     {
         $type = strtoupper(substr($value, -1));
-        $value = (integer) substr($value, 0, -1);
+        $value = substr($value, 0, -1);
 
         switch ($type) {
             case 'K' :
